Move sng to server/encoders/shellcode/sgn
