+++ server/encoders/shellcode/sgn/generate/main.go
