+++ server/encoders/shellcode/sgn/generate_files.go
