+++ server/encoders/shellcode/sgn/sgn.go
