+++ server/encoders/shellcode/sgn/testdata/windows-meterpreter-reverse-http.x86.bin
